backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-jose==3.3.0
email-validator==2.1.0.post1
aiohttp==3.9.1
motor==3.3.1
pytest==7.4.3
httpx==0.25.2
python-jose[cryptography]==3.3.0
boto3==1.34.14
huggingface-hub==0.19.4
ollama==0.1.6 